v2 - sam-med + gpt-4o/requirements_gpu.txt
# Enhanced V2 Pipeline Requirements with GPU Support

# Core dependencies (existing)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
matplotlib>=3.7.0
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0

# Azure ML dependencies
azure-ai-ml>=1.12.0
azure-identity>=1.12.0

# OpenAI API
openai>=1.0.0

# Optional GPU acceleration (significant performance improvement)
# Install these for GPU-accelerated clustering:
# cuml>=23.10.0  # RAPIDS cuML for GPU clustering (requires CUDA)
# cudf>=23.10.0  # RAPIDS cuDF for GPU dataframes

# UMAP for dimensionality reduction
umap-learn>=0.5.4

# CRF for spatial smoothing
pydensecrf>=1.0.2

# Note: For GPU support, you need:
# 1. NVIDIA GPU with CUDA capability >= 7.0
# 2. CUDA toolkit 11.2+ or 12.0+
# 3. Install RAPIDS cuML: conda install -c rapidsai -c conda-forge cuml=23.10 python=3.10 cudatoolkit=11.8
